+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/dto/StudentDTO.java
@@ -17,8 +17,6 @@ public class StudentDTO {
 
     private List<Collector> collectorList;
 
-    //TODO : Falta lista de Parents y Auth
-
     public StudentDTO(){
 
     }